+++ src/main/java/org/example/model/game/elements/Element.java
@@ -5,23 +5,23 @@ import org.example.model.game.Position;
 public class Element {
     private Position position;
 
-    public Element(int x, int y) {
+    public Element(float x, float y) {
         this.position = new Position(x, y);
     }
 
     public Position getPosition() {
         return this.position;
     }
-    public int getX(){
+    public float getX(){
         return position.getX();
     }
-    public int getY(){
+    public float getY(){
         return position.getY();
     }
-    public void setX(int x){
+    public void setX(float x){
         position.setX(x);
     }
-    public void setY(int y){
+    public void setY(float y){
         position.setY(y);
     }
